cmake: set install rpath explicitly on rocfft_rtc_helper and don't use rocm_install*
rocm_install* is opinionated about where to put a program like
rocfft_rtc_helper.  We want it in lib/rocfft/<version>, but rocm_install*
would want to put it in bin.

Previously we were hacking it with rocm_install with
$<TARGET_FILE:rocfft_rtc_helper>.  But then cmake only sees a filename, and
won't see INSTALL_RPATH information that's set on the target.